isRetryable fix
