--- SharedPackages/BrowserServicesKit/Sources/Common/OSStatus+Retryable.swift
@@ -1,38 +1,0 @@
-//
-//  OSStatus+Retryable.swift
-//
-//  Copyright © 2025 DuckDuckGo. All rights reserved.
-//
-//  Licensed under the Apache License, Version 2.0 (the "License");
-//  you may not use this file except in compliance with the License.
-//  You may obtain a copy of the License at
-//
-//  http://www.apache.org/licenses/LICENSE-2.0
-//
-//  Unless required by applicable law or agreed to in writing, software
-//  distributed under the License is distributed on an "AS IS" BASIS,
-//  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-//  See the License for the specific language governing permissions and
-//  limitations under the License.
-//
-// AI Generated 🤖
-
-import Foundation
-
-public extension OSStatus {
-
-    var isRetryable: Bool {
-        let retryableStatuses: Set<OSStatus> = [
-            errSecInteractionNotAllowed,
-            errSecNotAvailable,
-            errSecAuthFailed,
-            errSecAllocate,
-            errSecInternalError,
-            errSecInteractionRequired,
-            errSecIO,
-            errSecNotLoggedIn
-        ]
-
-        return retryableStatuses.contains(self)
-    }
-}
